requirements.txt
asgiref==3.4.1
Django==3.2.10
django-phonenumber-field==6.0.0
phonenumberslite==8.12.40
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
